Move PositionMonitor to position.rs
